SortableList: fix keyboard activation intercepting child button events #4371
Reorganized sortable list correctly.
